+++ src/Subscriptions/Storage/CacheStorageManager.php
@@ -30,7 +30,7 @@ class CacheStorageManager implements StoresSubscriptions
     /**
      * The time to live for items in the cache.
      */
-    protected int|null $ttl;
+    protected int|null $ttl = null;
 
     public function __construct(CacheFactory $cacheFactory, ConfigRepository $config)
     {