+++ CHANGELOG.md
@@ -40,6 +40,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Register all non-user namespaces through `RegisterDirectiveNamespaces` event https://github.com/nuwave/lighthouse/pull/2337
 - Guard default configuration options in `lighthouse.php` allows multiple guards https://github.com/nuwave/lighthouse/pull/2345
 - Make `@auth` and `@whereAuth` support multiple guards https://github.com/nuwave/lighthouse/pull/2345
+- Make `Illuminate\Http\Request` optional in `Nuwave\Lighthouse\Support\Contracts\GraphQLContext`
 
 ### Fixed